Restructored the storage of likes table again to support multiple endpoints
Storing the likes in an omeka-instance specific table didn't permit
overlapping order retention. For instance if a user liked an item from
site A, followed by a like from site B, then again from site A, all of
site A would be loaded first even though that wasn't the order the user
stored their likes. This is only essential for the default view of all
likes. To accomodate, secondary tables were eliminated and instead all
likes are registerd in a single table but have an instance-specific id.
So now a liked table may look like:

A-1
A-3
B-3
B-1
A-2